Fixed the location where the gradient was changed in the backward pass.
Since module backward hooks in PyTorch only really apply a hook onto the
last operation of the module, using the module backward hooks alone is
not enough. This commit introduces a 'Identity' function, which is
applied in a forward_pre_hook, and the backward_hook is attached to the
Identity grad_fn. The output will be saved to the Hook during the
Module's backward hook, which is a hook on the last operation of the
module and thus the true grad_output.