Be more careful with tensor references in Hook
The reference to hook.input seems to have retained, possibly due to
cyclic references. This fix now explicitly removes the references to
stored tensors when removing hooks, also the backward wrapper now only
gets a weak reference to the hook itself. This seems to fix the memory
leaks for now.