Add parameter control to mod_params and LinearHook
This commit adds control over the modified parameters in mod_params and
in LinearHooks by adding arguments param_keys and require_params.
Argument param_keys is a list of parameters that shall be modified, and
require_params is a boolean which, if True, raises a RuntimeError if
supplied parameter keys could not be found, and otherwise ignores
missing parameters.
The parameters modified by default were changed from weight and bias, to
all available ones.