Last epsilon of SmoothGrad set to zero
This sets the epsilon, added to the input in SmoothGrad, for the last
iteration to zero. With this, SmoothGrad with n_iter = 1 is equal to the
gradient (and not the gradient wrt. a shifted input). Additionally, the
model output for the SmoothGrad Attributor is now taken from the last
iteration, therefore saving the additional forward pass.