Canonizers: Add canonizers for torchvision densenets.
- Minor change in MergeBatchNormtoRight: remove unused variable
- Add DenseNetAdaptiveAvgPoolCanonizer: makes the last adaptive average pooling of torchvision densenets an explicit nn.module object
- Add ThreshReLUMergeBatchNorm: Canonizer to canonize BatchNorm -> ReLU -> Linear chains. Adds backwards and forward hooks to ReLU in order to turn it into ThreshReLU as defined in https://github.com/AlexBinder/LRP_Pytorch_Resnets_Densenet/blob/master/canonization_doc.pdf
- Add SequentialThreshCanonizer: a composite canonizer that applies DenseNetAvgPoolCanonizer, SequentialMergeBatchNorm, ThreshReLUMergeBatchNorm
- Add ThreshSequentialCanonizer: a composite canonizer that applies DenseNetAvgPoolCanonizer, ThreshReLUMergeBatchNorm, SequentialMergeBatchNorm

The last two canonizers are the recommended canonizers for torchvision DenseNet implementations. We need to apply the standard SequentialMergeBatchNorm to do away with the initial BN->Conv in the architecture. The two canonizers result in different implementations of the same function because in practice dense blocks have BN->ReLU->Conv->BN->ReLU->Conv which leaves the possibility of using the SequentialMergeBatchNorm inside the DenseBlocks if it is applied before. In practice, both canonizations get rid of the artifacts in the attribution maps. SequentialThreshCanonizer seems to be better quantitatively.